Bring mypy on src to zero and make it a check of CI
The last 51 findings, by kind:
- annotations that said less or more than the code does: the TK client returns iterables,
  a timeout is a float, the member helpers take any iterable, `RawSourceWriter` takes a
  `Store`, the citation extractor takes aliases without a target (it drops them itself);
- a loop variable `votes` that shadowed the `Votes` it came from (now `casts`), a
  `progress` declared twice, a filter on nodes without an id that the checker could not
  follow (a walrus in the comprehension keeps it);
- three SPARQL lambdas that bound loop variables through default arguments: named query
  functions and `partial`;
- python-arango types every call as its answer or a job: `schema._at_once` says that a
  plain connection answers at once, in one place;
- `requests` has no stubs installed: an override in `[tool.mypy]`, no new download.

One `type: ignore[override]` is left, with its reason: `retrieve tk-dossiers` has its own
options and does not use the fetch template of its base. `mypy` (>= 2.1, the version this
was checked with) is a dev dependency and a step of the CI test job. No behaviour changes.

+++ src/lawgraph/clients/base.py
@@ -64,7 +64,7 @@ class BaseClient:
         url: str,
         *,
         params: dict | None = None,
-        timeout: int = 30,
+        timeout: float = 30,
         stream: bool = False,
         headers: dict[str, str] | None = None,
     ) -> requests.Response:
@@ -97,7 +97,7 @@ class BaseClient:
         path: str,
         *,
         params: dict | None = None,
-        timeout: int = 30,
+        timeout: float = 30,
         retries: int = 5,
         backoff_factor: float = 2.0,
     ) -> requests.Response:
@@ -120,7 +120,7 @@ class BaseClient:
         url: str,
         *,
         params: dict | None = None,
-        timeout: int = 30,
+        timeout: float = 30,
         retries: int = 5,
         backoff_factor: float = 2.0,
         stream: bool = False,
@@ -187,7 +187,7 @@ class BaseClient:
         path: str,
         *,
         params: dict | None = None,
-        timeout: int = 30,
+        timeout: float = 30,
     ) -> dict[str, Any] | list[Any]:
         """Get JSON from the endpoint (with retry) and log item counts when present."""
         resp = self._get_raw_with_retry(path, params=params, timeout=timeout)
@@ -201,7 +201,7 @@ class BaseClient:
         path: str,
         *,
         params: dict | None = None,
-        timeout: int = 30,
+        timeout: float = 30,
     ) -> str:
         """Retrieve the text of the requested resource (with retry)."""
         resp = self._get_raw_with_retry(path, params=params, timeout=timeout)
@@ -212,7 +212,7 @@ class BaseClient:
         path: str,
         *,
         params: dict | None = None,
-        timeout: int = 30,
+        timeout: float = 30,
         result_key: str = "value",
         next_link_key: str | None = "@odata.nextLink",
     ) -> Iterator[dict[str, Any]]:
